requirements.txt
grpcio-tools
grpcio~=1.70.0
protobuf~=5.29.3
numpy
requests~=2.32.3
pillow~=11.1.0
prometheus_client~=0.21.1
prometheus_flask_exporter~=0.23.1
torch~=2.6.0
torchvision~=0.21.0
Flask~=3.1.0
prometheus-flask-exporter